requirements.txt
absl-py @ file:///C:/b/abs_14wchyjaia/croot/absl-py_1714140489450/work
aiohappyeyeballs @ file:///C:/b/abs_e9fm5mk6ak/croot/aiohappyeyeballs_1734469429350/work
aiohttp @ file:///C:/b/abs_d4_gdjffav/croot/aiohttp_1734687146788/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
altair==5.5.0
anyio==4.9.0
astunparse==1.6.3
async-timeout @ file:///C:/b/abs_d8yc2t_bxy/croot/async-timeout_1732663677668/work
attrs @ file:///C:/b/abs_89hmquz5ga/croot/attrs_1734533130810/work
blinker==1.9.0
brotlicffi @ file:///C:/b/abs_f4lixvzmxt/croot/brotlicffi_1736183286104/work
cachetools @ file:///C:/b/abs_fcvwpc5vh9/croot/cachetools_1738224687938/work
certifi @ file:///C:/b/abs_43o9lm92p9/croot/certifi_1750863864870/work/certifi
cffi @ file:///C:/b/abs_29_b57if3f/croot/cffi_1736184144340/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click @ file:///C:/b/abs_42rbjngfys/croot/click_1744271606983/work
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
cryptography @ file:///C:/b/abs_f4do8t8jfs/croot/cryptography_1694444424531/work
exceptiongroup==1.3.0
flatbuffers @ file:///C:/b/abs_7f2_85svxz/croot/python-flatbuffers_1722369110965/work
frozenlist @ file:///C:/b/abs_06ctmb1zeo/croot/frozenlist_1730903113463/work
gast @ file:///Users/ktietz/demo/mc3/conda-bld/gast_1628588903283/work
gitdb==4.0.12
GitPython==3.1.44
google-auth @ file:///C:/b/abs_86jbgkfwa2/croot/google-auth_1741277194305/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1617120569401/work
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
googletrans==4.0.2
grpcio @ file:///C:/b/abs_ccut8qek12/croot/grpc-suite_1681912611694/work
h11==0.16.0
h2==4.2.0
h5py @ file:///C:/b/abs_29off2a8q5/croot/h5py_1750863017528/work
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
importlib_metadata @ file:///C:/b/abs_9d0lp1ysmt/croot/importlib_metadata-suite_1732633510972/work
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras @ file:///C:/Users/builder/adipietro/mc3/tf210/conda-bld/keras_1669760570649/work/keras-2.10.0-py2.py3-none-any.whl
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
Markdown @ file:///C:/b/abs_f9hbi7znfr/croot/markdown_1746114643651/work
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
mpi4py @ file:///C:/b/abs_2c72fs0pz9/croot/mpi4py_1749458512296/work
multidict @ file:///C:/b/abs_19e3ubo2ew/croot/multidict_1730905504444/work
narwhals==1.45.0
numpy @ file:///C:/b/abs_c1ywpu18ar/croot/numpy_and_numpy_base_1708638681471/work/dist/numpy-1.26.4-cp39-cp39-win_amd64.whl#sha256=f284633067fde4bb078a09085a1c9e99a3218c129f5312b7b3d76c9340442a25
oauthlib @ file:///C:/b/abs_5e89byzts4/croot/oauthlib_1750775489850/work
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
packaging @ file:///C:/b/abs_3by6s2fa66/croot/packaging_1734472138782/work
pandas==2.3.0
pillow==11.3.0
propcache @ file:///C:/b/abs_57780xb0gk/croot/propcache_1744012740137/work
protobuf==3.20.3
pyarrow==20.0.0
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydeck==0.9.1
pyOpenSSL @ file:///C:/b/abs_08f38zyck4/croot/pyopenssl_1690225407403/work
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests @ file:///C:/b/abs_0b3h9d_724/croot/requests_1750426345668/work
requests-oauthlib @ file:///C:/b/abs_faw0a7_wag/croot/requests-oauthlib_1720615036938/work
rpds-py==0.26.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scipy @ file:///C:/Users/dev-admin/cr_scipy-feedstock/scipy_1733774324044/work/dist/scipy-1.13.1-cp39-cp39-win_amd64.whl#sha256=f7218697f964ae4372cdcb6f20a05db0d338ccc574c96e9dbf235ed6547dd462
six @ file:///C:/b/abs_149wuyuo1o/croot/six_1744271521515/work
smmap==5.0.2
sniffio==1.3.1
streamlit==1.46.1
tenacity==9.1.2
tensorboard @ file:///C:/Users/builder/adipietro/mc3/tf210/conda-bld/tensorboard_1669760968711/work/tensorboard-2.10.0-py3-none-any.whl
tensorboard-data-server @ file:///C:/b/abs_2fhvpo862s/croot/tensorboard-data-server_1670853600144/work/tensorboard_data_server-0.6.1-py3-none-any.whl
tensorboard-plugin-wit @ file:///C:/tf/b/tensorboard-plugin-wit_1660162132996/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorflow==2.10.0
tensorflow-estimator @ file:///C:/Users/builder/adipietro/mc3/tf210/conda-bld/tensorflow-estimator_1669761460695/work/tensorflow_estimator-2.10.0-py2.py3-none-any.whl
termcolor @ file:///C:/b/abs_16qe7jhw7n/croot/termcolor_1668084642458/work
toml==0.10.2
tornado==6.5.1
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
tzdata==2025.2
urllib3 @ file:///C:/b/abs_393g1ayee_/croot/urllib3_1750775481226/work
watchdog==6.0.0
Werkzeug @ file:///C:/b/abs_c7bupijx2_/croot/werkzeug_1737448709012/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wrapt @ file:///C:/b/abs_68gzx33lzh/croot/wrapt_1736543196433/work
yarl @ file:///C:/b/abs_281qby2vim/croot/yarl_1732546854547/work
zipp @ file:///C:/b/abs_2fcpcg89my/croot/zipp_1732630746332/work
